change int to double for ingredient amount param

+++ src/main/java/io/catalyte/demo/ingredient/Ingredient.java
@@ -13,7 +13,7 @@ public class Ingredient {
     private Boolean active;
     private String name;
     private BigDecimal purchasingCost;
-    private int amount;
+    private double amount;
     private String unitOfMeasure;
 
     @ElementCollection
@@ -22,7 +22,7 @@ public class Ingredient {
     public Ingredient() {}
 
     public Ingredient(int id, Boolean active, String name, BigDecimal purchasingCost,
-                      int amount, String unitOfMeasure, List<String> allergens) {
+                      double amount, String unitOfMeasure, List<String> allergens) {
         this.id = id;
         this.active = active;
         this.name = name;
@@ -64,11 +64,11 @@ public class Ingredient {
         this.purchasingCost = purchasingCost;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
